+++ source/app/rendering/camera.cpp
@@ -206,9 +206,11 @@ bool Camera::unproject(int x, int y, int z, QVector3D& result) const
     y = _viewportHeight - y;
 
     QVector4D normalisedCoordinates;
-    normalisedCoordinates.setX((x / static_cast<float>(_viewportWidth)) * 2.0f - 1.0f);
-    normalisedCoordinates.setY((y / static_cast<float>(_viewportHeight)) * 2.0f - 1.0f);
-    normalisedCoordinates.setZ(2.0f * z - 1.0f);
+    normalisedCoordinates.setX((static_cast<float>(x) /
+        static_cast<float>(_viewportWidth)) * 2.0f - 1.0f);
+    normalisedCoordinates.setY((static_cast<float>(y) /
+        static_cast<float>(_viewportHeight)) * 2.0f - 1.0f);
+    normalisedCoordinates.setZ(2.0f * static_cast<float>(z) - 1.0f);
     normalisedCoordinates.setW(1.0f);
 
     QVector4D o = m * normalisedCoordinates;
