+++ source/app/rendering/primitives/arrow.h
@@ -25,7 +25,7 @@ public:
 
 public slots:
     void setRadius(float radius) { _radius = radius; }
-    void setLength(int length) { _length = length; }
+    void setLength(float length) { _length = length; }
     void setSlices(int slices) { _slices = slices; }
 
     void create(QOpenGLShaderProgram& shader);
