+++ source/app/updates/updater.cpp
@@ -211,7 +211,7 @@ void Updater::downloadUpdate(QNetworkReply* reply)
         connect(_reply, &QNetworkReply::downloadProgress, [this]
         (qint64 bytesReceived, qint64 bytesTotal)
         {
-            _progress = (bytesReceived * 100) / bytesTotal;
+            _progress = static_cast<int>((bytesReceived * 100u) / bytesTotal);
             emit progressChanged();
         });
